Danny, Meghna, Grace, please review.
These are initial versions of a few key component interfaces.

manager/Manager.java - the interface between the servlet layer and the entire connector manager.
persist/ConnectorConfigStore.java - the interface between the instantiator and persistent store.
persist/ConnectorStateStore.java - the interface between traversal methods and persistent store.
traversal/TraversalMethod.java - the interface between the controller and traversal methods.

These are all version 0 - it's understood that we will change them as necessary.  But I need most of these to check in an initial version of the traversal logic, and I want for us all to have something concrete to work from.