Danny, please review.
SimpleConfigurer.java - a basic implementation of the Configurer interface for the simple situation where the config is kjust a list of keys and there is no I18N.  Implementors may use it directly or just as a model.  (Passing question - perhaps it might be better to have a separate package for this sort of thing: maybe c.g.e.connector.spi.sdk?)

SimpleConfigurerTest.java - test for SimpleConfigurer.java

JsonEntrySet.java, JsonObjectAsMap.java - convenience classes for writing tests.