Fix Issue 75 - Http 400 Error Thrown For Correct Feeds (When Using Multiple Connectors)
This Was Related To The Docpusher.Take() Method Not Being Thread-Safe.  I Have
Removed The Use Of Several Instance Fields (Xmldata, Datasource, And Feedtype)
From The Method And Moved The New Feedlogrecord Into A Threadlocal.  The
Feedconnection.Senddata() Seems To Be Thread-Safe Since It Creates A New
Connection Each Time So I Have Left That As It Is. 

It Should Be Noted That The Retrieval Gsaresponse Instance Field Is Not
Thread-Safe, However, It Is Only Used From Tests.

Tested With 2 Connectors Each Feeding 600 Documents And No Problems.