Fix Issue 75 - Http 400 Error Thrown For Correct Feeds (when using multiple connectors)
This was related to the DocPusher.take() method not being thread-safe.  I have
removed the use of several instance fields (xmlData, dataSource, and feedType)
from the method and moved the new feedLogRecord into a ThreadLocal.

I have also removed the lazy evaulation and potential DCL problem from the
GsaFeedConnection.

It should be noted that the retrieval of the GsaResponse instance field is not
thread-safe, however, it is only used from tests.

Tested with 2 connectors each feeding 600 documents and no problems.