[Copy NullFeedConnection from patch branch to trunk.]
The NullFeedConnection replaces GsaFeedConnection.  It consumes
the feed data, but rather than sending it to the GSA, it simply
discards the data.  This can be used to identify or eliminate
the GSA as problem point in the system.  If replacing the
GsaFeedConnection fails to correct the performance issues, then
the problem lies elsewhere.

Using the NullFeedConnection is also a benefit for Connector
developers.  Traversal behaviour and feed content can be observed
without requiring direct access to a GSA during development.
This is especially useful when used in conjunction with Feed logging
or teedFeedFile logging to examine what data would have been fed
to the GSA.

Replacing GsaFeedConnection with NullFeedConnection requires modifying
${TOMCAT_HOME}/webapps/connector-manager/WEB-INF/applicationContext.xml
where ${TOMCAT_HOME} represents the Connector Manager Tomcat installation.

1) Make a backup of copy of applicationContext.xml

2) Edit applicationContext.xml, and locate the FeedConnection bean definition:

   <bean id="FeedConnection"
       class="com.google.enterprise.connector.pusher.GsaFeedConnection">
       <constructor-arg index="0" type="java.lang.String"
            value="${gsa.feed.host}" />
       <constructor-arg index="1" type="int" value="${gsa.feed.port}" />
   </bean>

3) Replace that bean definition with:

   <bean id="FeedConnection"
       class="com.google.enterprise.connector.pusher.NullFeedConnection">
   </bean>

4) Restart Tomcat


Since the NullFeedConnection does not actually send any content
to the GSA, you will need to reindex your data once you restore
the GsaFeedConnection, as any documents that were sent to the
NullFeedConnection never actually got indexed.

You can avoid reindexing by either:
  A) Troubleshooting with a completely separate Tomcat/ConnectorManager
       installation.
or
  B) Suspending indexing, then troubleshooting with a separate
     Connector instance:
     1) Suspend indexing by setting your Connector schedule to 1am - 1am.
     2) Wait 10 minutes.
     3) Shutdown Tomcat.
     4) Replace GsaFeedConnection with NullFeedConnection as decribed above.
     5) Restart Tomcat.
     6) Create a new Connector instance with the same configuration as your
        existing connector instance, but a different name.

     Observe the traversal/feed behaviour -- Achieve enlightenment.
     When you are ready to restore the GsaFeedConnection:
     7) Delete the Connector instance created in the previous step.
     8) Shutdown Tomcat.
     9) Replace applicationContext.xml with the backup you made.
    10) Restart Tomcat.
    11) Resume indexing by setting your Connector's schedule to 12am-12am.