This sweeping change cleans up some whitespace-related issues in the code base.
According to Google style guides, the source code should use
spaces, not tabs, for indentation, and lines should have no
trailing whitespace.  These rules are easy to set up in emacs,
however when I do so, simple changes have many diffs, because
the codebase is filled with tabs and trailing whitespace that
get "corrected" when I edit a file.

Similarly, I occasionally encountered files that had both
Unix line-endings and MS-DOS line-endings.  This too annoyed
emacs, since it wouldn't edit the file in DOS mode because
a handful of lines did not have a CR.

Rather than trying to isolated this noise from my real changes,
I simply made a sweep of the codebase, fixing up a bunch of
whitespace and formatting issues.  No functionality was changed*.
[*Actually one unit test was modified, because its test file
changed in length.]

The changes are as follows:
 - Untabify - tabs converted to spaces
 - Strip trailing whitespace on lines
 - Reformating indents that were unbalanced by poor tabify (mostly xml files)
 - Insert Copyright/License comment block on files that were missing them.
 - Reformat some lines do to 80 column guidlines (although not many)
 - Normalize line endings to Unix-style NL. (In anticipation of eol-style=native property).