This change attempts to deal with failures of ThreadPoolTest on a slow machine. I have not seen these tests fail on my machine.
The failure this fix attempts to address is in 
 testTimeToLiveWithHungBatch

The test
 1) starts a task with 100 milliseconds to live
 2) confirms the task is running by reading from a qaueue the task
    writes to
 3) verifies the task has been canceled
 
From the failure log it appears the thask does perform step 2. I
think this is because the task is cancelled before it gets the chance. Tests
seem to be taking 50 to 100 times longer on this machine than on my development
machine. This change increases the time before the cancel from 100 ms
to 250 ms to give the task more time to run.