Moving the call to increment the counter and check for batchHint to a point before the call to pusher.take() so it will always be incremented when a document is pulled from the DocumentList.  Currently it is only incremented after a successful call to take().  Any exceptions thrown by take() would cause the counter and check for batch hint to be skipped.  This could result in more documents being pulled from the host than reported and the host load limit could be exceeded.
Note, it's still the case that if the DocumentList.nextDocument() throws a
RespositoryDocumentException that call to nextDocument() will not count against
the batchHint.  A summary of the new contract is once a document is returned
from nextDocument() it will count against the batchHint.  It used to be the case
that a document could be returned and it would not be counted if an exception
was thrown while extracting the docid or trying to push the document to the feed
host.

Tests Passed:
  - All Unit Tests.

Change Log:
  M java/com/google/enterprise/connector/traversal/QueryTraverser.java:
    - Moved the check for batchHint to the top of the loop rather than the
      bottom.