Fix for issue 175.
The root cause of this issue is that when the connector manager starts up and loads a ConnectorCoordinatorImpl
for each connector instance the code was not correctly initializing the typeInfo field.

When a connector instance is deleted, the ConnectorCoordinatorImpl
is cleared but not removed from SpringInstantiator.coordinatorMap. The reason for not removing the 
ConnectorCoordinatorImpl is to assure proper coordinate between concurrent deletions and updates or
of a connector instance. The check if the ConnectorCoordinatorImpl has been cleared is run by
calling getInstanceInfo. This change also calls getInstanceInfo first to run the check before
attempting to access a possibly cleared ConnectorCoordinatorImpl.