Issue 106: Handling of large or empty document content.
This change forces document content that exceeds the maxDocumentSize
to be discard and replaced with the default placeholder content
(the google:title value or a space).  This mirrors the behaviour
of Connectors that implement TraversalContextAware and choose
to no supply a content stream for documents that exceed the
TraversalContext.maxDocumentSize.

As a consequence of this change, I also cleared up a TODO:
Documents with non-null, but empty, content streams also
provide the default placeholder content.

The mechanics of the change are simple.  DocPusher's
BigDocumentFilterInputStream now takes the XmlFeed
(so it can reset to the begining of the content section),
and an alternate content InputStream, in addition to
the contentstream.  If the content read exceeds the
maximum document size, BigDocumentFilterInputStream
resets the feed back to the start of the content,
then switches to streaming the alternate content.