Issue 187 - Fix determineBatchHint log messages once per second.
TraversalScheduler tries to schedule a batch for the connector
once per second.  But it doesn't know if the connector is already
processing a batch.

Some analysis reveals the problem is related to asynchronous
worker threads.  The check for "is this connector already
processing a batch?" was moved down into
ConnectorCoordinatorImpl.startBatch(), after the call to
determineBatchHint().  So for every second that the
connector is processing a batch, a determineBatchHint()
message gets added to the log.

The quick and easy (read ugly) way to fix this was to create
a subclass of BatchSize that lazily evaluates the BatchSize,
waiting until one of the getters are called.  This maintains
the calling convention of TraversalScheduler passing the
BatchSize down to the ConnectorCoordinator, but delays calling
HostLoadManager.determineBatchSize() until after the Coordinator
as determined that it is OK to start a new batch.

The real way to fix this is to move host load management for
a connector instance into the ConnectorCoordinator.  Once that
is done, the BatchSize does not need to get passed down from
the Scheduler to the Coordinator.  See Issue 189.