Fix ConnectorCoordinatorImpl.restartConnectorTraversal()
ConnectorCoordinator maintains a local cache of the Schedule object.
It also maintains the timer that delays traversals after an error or
after reaching the end of the repository.

ConnectorCoordinator's handling for restartTraversal would bypass
it's own code for setting the schedule.  In doing so, it would leave
the cached Schedule stale.  It also didn't reset the traversal delay
timer, so the restart might not actually start for 15 minutes.

This change fixes both of those problems.  When modifying the schedule
(to reset a run-once schedule), the code now calls the local setSchedule()
method, which resets the cached schedule and the delay timer.  It also
resets the timer in the cases where the schedule would not be modified.

This bug was caught during code inspection for an unrelated task.

Changed files:
-------------
M projects/connector-manager/source/java/com/google/enterprise/connector/instantiator/ConnectorCoordinatorImpl.java