The main goal of this CL is to change the build of the spi-jar so that it has no dependencies on connector-manager code outside itself. And helps by providing allowing us to unset "sourcepath" which prevents the compiler from automatically following import dependencies and compiling them; see: http://ant.apache.org/manual/Tasks/javac.html#srcdirnote
The main build.xml file is changed:
* to use a new target directory for spi classes
* to build the spi jar only from those classes
* to include **/spi/**, **/spiimpl/** and **/util/** in the spi
* to build other targets from built jars, rather than classes directories

The code changes are the result of finding (as Brett predicted) some existing
dependencies from the SPI to the CM, in particular, EofFilterInputStream. I
moved that to util.

Also, I moved some testing stuff from the javatests tree c.g.e.c.util.diffing to
the java tree c.g.e.c.util.testing.diffing, as we discussed.

Finally, I simplified the builds of testconnectors A and B - which was another
change that had been sitting around in my workspace for a while.

http://codereview.appspot.com/2203041/