The goal of this cl is to get the 2.6.x branch back together with 2.6.4 without losing any 2.6.x history.
I created this cl as follows: 
1. I checked out 2.6.x and 2.6.4 into separate directories
2. I created git repositories in the same directories
3. I cloned the git repos in two other separate directories
4. I erased everything from the 2.6.x repo
5. I copied everything from the 2.6.4 repo to 2.6.x
6. I did a git commit
7. I pushed the git changes from the clone to the 2.6.x repo I created in 1
8. I did an svn revert on projects/version.properties
9. I ran upload.py

I passed via git because, unlike svn, it does not keep little directories of its
own in each sub-directory and it doesn;t get confused if identical files are
written on top of each other. Maybe there was a better way to do this - but it
worked for me.

http://codereview.appspot.com/2254042/