Log more crawl activities while crawling in diffing connector library Added an interface called "CrawlActivityLogger" with method for each crawl event ( crawl start, end, new doc, changed doc, deleted doc etc)
Added a default behaviour "DefaultCrawlActivityLogger" which just puts in a log
statement for each event and at the "end crawl" event logs a summary 

For this, added a new configuration in the spring file to take an instance of
"CrawlActitiyLogger" as a constructor argument for "ChangeQueue" class.
Intention is to expose this as a configuration so that people (if required) can
replace the default behavior with their own custom Crawl Activity loggers. 

CL: http://codereview.appspot.com/4427054/