Remove caching of feedConnection features in DocPusher.
GsaFeedConnection and DocPusher cached certain GSA features.
These caches were not invalidated if the user switched GSAs,
specifically GSAs of different versions with different features.

The caching in GsaFeedConnection was fixed in r3144.

This fixes the caching in DocPusher.  To simplify the fix,
the construction of the content stream was moved from DocPusher
to XmlFeed.

Code Review:
http://codereview.appspot.com/6849110