Fix bug 10248294: Connector Manager GetConfig should be a downloadable file
This change reverts the XML version of the GetConfig servlet on the trunk
intially made in r2568.

The XML version was intended for HA deployments, with connector configurations
stored in a database - none of which ever really happened.

The XML version also did not handle external state files like those in the
Sharepoint and Diffing connectors.

This change restores the old version of the GetConfig servlet that ZIPs
up the web application WEB-INF directory and returns it.  Specifically,
this version of the GetConfig servlet is from the 3.0.x branch revision
r2967.  See also a similar revert made in r2803.

Because there was not a unit test for the old version of GetConfig,
I could not revert GetConfigTest.  I therefore disabled the test with
a TODO to fix it.

Code Review:
http://codereview.appspot.com/12783044