Fix deprecation warnings.
Suppressing deprecation warnings on a method that used
ServletUtil.XMLTAG_CONNECTOR_SCHEDULE was not working, so I copied the
constant to a local constant and suppressed warnings on the copy.

Removing the import of SpiConstants.RoleType and qualifying it fixed a
strange javac warning (yes, in the wrong file, which did not refer to
RoleType):

AlternateContentFilterInputStream.java:51: warning: [deprecation]
RoleType in SpiConstants has been deprecated
   * has null or empty content, or if the feed item has excessively large
                                   ^

Removing the use of ServletUtil.XMLTAG_CONNECTOR_SCHEDULE completely
and fixing the GData JAR file warnings is left for future changes.

Code review: https://codereview.appspot.com/28230043/