Merge changes to ConnectorFields from the file system connector.
The file system connector was the original, I think, and then it was
copied to the LDAP connector, and then here to the CM. The copy in the
file system connector was later modified to XML encode the input label
in AbstractField, and to add a default value to SingleLineField.

I added XML encoding for the label and also the optional message. I
added the default value as a setDefaultValue method, rather than a
constructor argument, for better backward compatibility and
parallelism with setBoldLabel.

Other changes:

* Fixed a typo in a parameter name in AbstractField.getSnippet.

Code review: https://codereview.appspot.com/48860043/