Set authmethod="none" for public documents (b/15398588).
Previously, we did not set the attribute for public documents, when
PROPNAME_ISPUBLIC was missing or set to true. Before GSA 7.0, the
default authmethod value was always "none". Now the default is to
preserve the existing value if a document was already indexed. If a
document changes from secure to public, without the explicit
authmethod attribute the previous value of secure is preserved.

If PROPNAME_ISPUBLIC was false, then PROPNAME_AUTHMETHOD was used for
the authmethod attribute, and if PROPNAME_ISPUBLIC was missing or
true, then PROPNAME_AUTHMETHOD was ignored. Now the more secure of the
two values is used and a warning is logged.

Other changes:

* Fix the documentation about ACLs and PROPNAME_ISPUBLIC. The doc said
  that ACLs are ignored if PROPNAME_ISPUBLIC is missing or set to true,
  but this was never true. Even the original ACL CL (f40a23a7)
  modified DocPusherTest.testTakeIsPublic to include ACLs with
  google:ispublic set to true.

* Add tests for ValueImpl.toBoolean with all Value types.

* Copy the TestUtil class from the Notes connector, renamed to
  Logging, to capture and test log messages.

Also manually tested all combinations and log messages.