Fix svn-branch target for branches with tags.
Git branches may have tags on earlier commits, such as release tags on
the release branches, and in that case "git describe --tags" will
produce something "<branch>-<n>-<commit>" rather than the desired
"<branch>". Adding "--exact-match" will only match tags on the HEAD.

This is still a change in behavior, because it we have the branch
checked out, but it has been tagged, then the svn.branch property will
now be the tag rather than the branch. That could be taken as a
feature, since we will have have to checkout the tag separately, but
can simply build from the branch once the tag is created.

Code review: https://codereview.appspot.com/254730043/