Always feed ACLs in acl elements rather than as metadata.
The feed.disabled.inherited.acls property still determines the return
value for TraversalContext.supportsInheritedAcls, but internally we
assume that we can always feed acl elements (including inherited ACLs)
to the GSA. This works with GSA 7.0 and later.

The rationale is to avoid the use of legacy ACLs in metadata moving
forward so that they can be deprecated on the GSA.

For posterity, the comment on ProductionManager.setFeedConnection was a
copy-and-pasto from the same method in GetDocumentContent, and it had
nothing to do with the security header. Instead, as now noted in the
comment on the ProductionManager copy, the FeedConnection was used to
determine whether feeds supported inherited ACLs.

Code review: https://codereview.appspot.com/253800043/